refactor: 메서드 위치 수정
